qianjunm

+++ app/controllers/sales_controller.rb
@@ -2,9 +2,11 @@
 class SalesController < ApplicationController    #营销管理 -- 活动
   before_filter :sign?
   layout 'sale'
-
+  require "toPinyin"
+  
   #活动列表
   def index
+    p " 活动列表".pinyin
     @sales=Sale.paginate_by_sql("select s.id,name,s.store_id,s.started_at,s.everycar_times,s.disc_time_types,s.ended_at,s.code,s.status
     from sales s where s.store_id=#{params[:store_id]} and s.status !=#{Sale::STATUS[:DESTROY]} order by s.created_at desc ", :page => params[:page], :per_page => Constant::PER_PAGE)
     @orders = Order.select("sale_id,count(id) num").where(:store_id=>params[:store_id],:status=>[Order::STATUS[:BEEN_PAYMENT],Order::STATUS[:FINISHED]]).
@@ -105,9 +107,10 @@ class SalesController < ApplicationController    #营销管理 -- 活动
   
   def upload_stream(img_url,dirs,img_code=nil)
     path = Constant::LOCAL_DIR + dirs.join("/")
-    FileUtils.remove_dir path
-    FileUtils.mkdir_p  path
-    dirs << "#{dirs[2]}img."+ img_url.original_filename.split(".").reverse[0]
+    FileUtils.remove_dir path if  File.directory? path
+    FileUtils.mkdir_p  path unless  File.directory? path
+    filename = img_url.original_filename.split(".")
+    dirs << "#{filename[0].pinyin.push(dirs[2]).join("")}."+ filename.reverse[0]
     path = Constant::LOCAL_DIR + dirs.join("/")
     File.open(path, "wb")  {|f|f.write(img_url.read);}
     return "/"+dirs.join("/")
